BIG UPDATES TO EXECUTESECURITYTESTS
I changed up a bunch of stuff and finally have something so we can figure out what our strategy for changing values for parameters and actually testing is. I also added default constructors to range so we don't have to put those psuedo ranges in. I want to explain later everything in person.